#32 Open resource from non-imported project with internal editor
Non-UI resources moved to a separate branch
