+++ bundles/org.platformio.eclipse.ide.home.core/META-INF/MANIFEST.MF
@@ -11,5 +11,5 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.20.100",
  org.platformio.eclipse.ide.home;bundle-version="0.1.0",
  org.platformio.eclipse.ide.installer;bundle-version="0.1.0"
 Bundle-ActivationPolicy: lazy
-Service-Component: OSGI-INF/org.platformio.eclipse.ide.home.internal.core.HomeAccess.xml
 Export-Package: org.platformio.eclipse.ide.home.core
+Service-Component: OSGI-INF/org.platformio.eclipse.ide.home.internal.core.HomeAccess.xml
