+++ bundles/org.platformio.eclipse.ide.home.ui/META-INF/MANIFEST.MF
@@ -17,7 +17,8 @@ Require-Bundle: org.eclipse.core.resources;bundle-version="0.0.0",
  org.eclipse.ui;bundle-version="0.0.0",
  org.platformio.eclipse.ide.home.core;bundle-version="0.1.0",
  org.platformio.eclipse.ide.home;bundle-version="0.0.0",
- org.platformio.eclipse.ide.workbench;bundle-version="0.1.0"
+ org.platformio.eclipse.ide.workbench;bundle-version="0.1.0",
+ org.platformio.eclipse.ide.workspace;bundle-version="0.0.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
 Automatic-Module-Name: org.platformio.eclipse.ide.home.ui
