+++ bundles/org.platformio.eclipse.ide.workbench/META-INF/MANIFEST.MF
@@ -13,12 +13,12 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0",
  org.eclipse.core.resources;bundle-version="0.0.0",
  org.eclipse.osgi.services;bundle-version="0.0.0",
  org.eclipse.ui.ide;bundle-version="0.0.0",
- org.eclipse.ui.workbench.texteditor;bundle-version="3.16.0",
- org.eclipse.jface.text;bundle-version="3.17.0",
- org.eclipse.ui.navigator.resources;bundle-version="3.8.0",
- org.eclipse.ui.navigator;bundle-version="3.10.0",
- org.eclipse.ui.views;bundle-version="3.11.0"
+ org.eclipse.ui.workbench.texteditor;bundle-version="0.0.0",
+ org.eclipse.jface.text;bundle-version="0.0.0",
+ org.eclipse.ui.navigator.resources;bundle-version="0.0.0",
+ org.eclipse.ui.navigator;bundle-version="0.0.0",
+ org.eclipse.ui.views;bundle-version="0.0.0",
+ org.platformio.eclipse.ide.workspace;bundle-version="0.0.0"
 Bundle-ActivationPolicy: lazy
-Service-Component: OSGI-INF/org.platformio.eclipse.ide.workbench.ImportProject.xml,
- OSGI-INF/org.platformio.eclipse.ide.workbench.OpenFileHandler.xml
 Export-Package: org.platformio.eclipse.ide.workbench
+Service-Component: OSGI-INF/org.platformio.eclipse.ide.workbench.OpenFileHandler.xml
