--- bundles/org.platformio.eclipse.ide.workbench/OSGI-INF/org.platformio.eclipse.ide.workbench.ImportProject.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.platformio.eclipse.ide.workbench.ImportProject">
-   <service>
-      <provide interface="org.platformio.eclipse.ide.home.net.IDECommand"/>
-   </service>
-   <implementation class="org.platformio.eclipse.ide.workbench.ImportProject"/>
-</scr:component>
