+++ bundles/org.platformio.eclipse.ide.workbench/src/org/platformio/eclipse/ide/workbench/OpenFileHandler.java
@@ -20,8 +20,6 @@
  *******************************************************************************/
 package org.platformio.eclipse.ide.workbench;
 
-import java.util.Optional;
-
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IStatus;
@@ -37,31 +35,30 @@ import org.eclipse.ui.ide.IDE;
 import org.eclipse.ui.texteditor.ITextEditor;
 import org.osgi.service.component.annotations.Component;
 import org.platformio.eclipse.ide.home.net.IDECommand;
+import org.platformio.eclipse.ide.workspace.OpenFile;
 
 import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
 
 @Component
 public final class OpenFileHandler implements IDECommand {
 
 	@Override
 	public void execute(JsonElement element) {
+		OpenEditorData data = new OpenEditorData(element);
+		new OpenFile(data.path()).get().ifPresent(file -> open(file, data.line()));
+	}
+
+	private void open(IFile file, int line) {
 		IWorkbench workbench = PlatformUI.getWorkbench();
 		workbench.getDisplay().syncExec(() -> {
 			try {
-				JsonObject result = element.getAsJsonObject().get("params").getAsJsonObject(); //$NON-NLS-1$
-				String path = result.get("path").getAsString(); //$NON-NLS-1$
-				int line = result.get("line").getAsInt(); //$NON-NLS-1$
-				Optional<IFile> file = new OpenFile(path).get();
-				if (file.isPresent()) {
-					IWorkbenchWindow window = window(workbench);
-					IEditorPart editor = IDE.openEditor(window.getActivePage(), file.get());
-					if (editor instanceof ITextEditor) {
-						IRegion region = ((ITextEditor) editor).getDocumentProvider()
-								.getDocument(editor.getEditorInput()).getLineInformation(line - 1);
-						((ITextEditor) editor).getSelectionProvider()
-								.setSelection(new TextSelection(region.getOffset(), region.getLength()));
-					}
+				IWorkbenchWindow window = window(workbench);
+				IEditorPart editor = IDE.openEditor(window.getActivePage(), file);
+				if (editor instanceof ITextEditor) {
+					IRegion region = ((ITextEditor) editor).getDocumentProvider().getDocument(editor.getEditorInput())
+							.getLineInformation(line - 1);
+					((ITextEditor) editor).getSelectionProvider()
+							.setSelection(new TextSelection(region.getOffset(), region.getLength()));
 				}
 			} catch (Exception e) {
 				Platform.getLog(getClass()).info(e.toString());
