+++ bundles/org.platformio.eclipse.ide.workspace/.settings/org.eclipse.pde.ds.annotations.prefs
@@ -1,0 +1,8 @@
+classpath=true
+dsVersion=V1_3
+eclipse.preferences.version=1
+enabled=true
+generateBundleActivationPolicyLazy=true
+path=OSGI-INF
+validationErrorLevel=error
+validationErrorLevel.missingImplicitUnbindMethod=error
