+++ bundles/org.platformio.eclipse.ide.workspace/OSGI-INF/org.platformio.eclipse.ide.workspace.ImportProject.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.platformio.eclipse.ide.workspace.ImportProject">
+   <service>
+      <provide interface="org.platformio.eclipse.ide.home.net.IDECommand"/>
+   </service>
+   <implementation class="org.platformio.eclipse.ide.workspace.ImportProject"/>
+</scr:component>
