+++ features/org.platformio.eclipse.ide.feature/feature.xml
@@ -117,5 +117,12 @@
          install-size="0"
          version="0.0.0"
          unpack="false"/>
+   
+   <plugin
+         id="org.platformio.eclipse.ide.workspace"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
 
 </feature>
