+++ releng/org.platformio.eclipse.ide.aggregator/pom.xml
@@ -43,6 +43,7 @@
 		<module>..\..\bundles\org.platformio.eclipse.ide.home.core</module>
 		<module>..\..\bundles\org.platformio.eclipse.ide.home.ui</module>
 		<module>..\..\bundles\org.platformio.eclipse.ide.workbench</module>
+		<module>..\..\bundles\org.platformio.eclipse.ide.workspace</module>
 		<module>..\..\features\org.platformio.eclipse.ide.feature</module>
 
 		<module>..\..\releng\org.platformio.eclipse.ide.repository</module>
